Simplify the sources repo
